+++ src/components/map/layers/CartoDbLayer.vue
@@ -17,7 +17,7 @@ watch(map, (val) => {
     if (!tileLayer) {
         tileLayer = new TileLayer({
             source: new XYZ({
-                attributions: '© <a href="https://navigraph.com/" target="_blank">Navigraph</a> © <a href="https://openlayers.org/" target="_blank">OpenLayers</a> © <a href="http://cartodb.com/attributions" target="_blank">CartoDB</a>',
+                attributions: '© <a href="/about" target="_blank">3rd Party Projects</a> © <a href="http://cartodb.com/attributions" target="_blank">CartoDB</a>',
                 url: 'https://cartodb-basemaps-a.global.ssl.fastly.net/dark_all/{z}/{x}/{y}.png',
                 wrapX: true,
             }),