+++ src/composables/data.ts
@@ -2,9 +2,11 @@ import type { VatDataVersions } from '~/types/data';
 import type { VatSpyAPIData } from '~/types/data/vatspy';
 import type { VatsimLiveData } from '~/types/data/vatsim';
 import type { Ref } from 'vue';
+import type { SimAwareAPIData } from '~/utils/backend/storage';
 
 const versions = ref<null | VatDataVersions>(null);
 const vatspy = shallowRef<VatSpyAPIData>();
+const simaware = shallowRef<SimAwareAPIData>();
 
 type Data = {
     [K in keyof VatsimLiveData]: Ref<VatsimLiveData[K] extends Array<any> ? VatsimLiveData[K] : (VatsimLiveData[K] | null)>
@@ -25,6 +27,7 @@ const data:Data = {
 
 const vatsim = {
     data,
+    versions: ref<VatDataVersions['vatsim'] | null>(null),
     updateTimestamp: ref(''),
 };
 
@@ -33,6 +36,7 @@ export function useDataStore() {
         versions,
         vatspy,
         vatsim,
+        simaware,
     };
 }