+++ src/store/index.ts
@@ -1,6 +1,7 @@
 import { defineStore } from 'pinia';
 import type { FullUser } from '~/utils/backend/user';
 import type { UserLocalSettings } from '~/types/map';
+import type { ThemesList } from '~/modules/styles';
 
 export const useStore = defineStore('index', {
     state: () => ({
@@ -8,6 +9,7 @@ export const useStore = defineStore('index', {
 
         user: null as null | FullUser,
         version: '',
+        theme: 'default' as ThemesList,
         localSettings: {} as UserLocalSettings,
     }),
 });